Optimization and bug fixes
